requirements.txt
streamlit
pandas
openai
openpyxl
python-dotenv
reportlab
altair
requests